+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project changes will be shown here.
 - Issue 926 - Pending Reopen New SR doesn't delete the pending one.
 - Issue 904 - Renewal Report Unit numbers not matching
 - Issue 892 - Lead Status Layout Change
+- Issue 822 - Combining SRs for Sales Order ordering issue
 
 ## [v4.1.7] - 2020-06-15
 ### Added